\name{send_message_card}
\alias{send_message_card}
\title{
  Send Message Card
}
\description{
  Send Message Card to Teams Channel
  
}
\usage{
send_message_card(wh, msg, ...)
}
\arguments{
  \item{wh}{
    webhook
}
  \item{msg}{
    message (a string)
}
  \item{\dots}{
    
}
}
\details{
  Experimental.
}
\value{
  the return of \code{\link[httr]{POST}}
}
%% \references{
%%   %% ~put references to the literature/web site here ~
%% }
\author{
  Enrico Schumann
}
%% \seealso{
%% %% ~~objects to See Also as \code{\link{help}}, ~~~
%% }
\examples{
### to be added
}
